fingerprint-bridge/requirements.txt
pyserial==3.5
requests==2.31.0
Pillow==10.2.0
